+++ scripts/run_demo.sh
@@ -47,6 +47,9 @@ if [ $wait -gt 60 ]; then
   exit -1
 fi
 
+docker run --env MONGO_DB="mongodb://mongo:27017/graphql" --name=populateDB --network=anubis_envoymesh orchestracities/configuration-api node main/mongo/populateDB.js
+docker rm -f populateDB
+
 echo "Setting up tenant Tenant1..."
 curl -s -i -X 'POST' \
   'http://127.0.0.1:8085/v1/tenants/' \
@@ -171,4 +174,18 @@ curl -s -i -X 'POST' \
 "agent": ["acl:AuthenticatedAgent"]
 }'
 
-echo "Demo deployed!"
+if [[ $1 == "silent" ]]; then
+  echo "Demo deployed!"
+else
+  echo "Your browser will open at: http://localhost:3000"
+  echo "User: admin / Password: admin"
+  if [ "$(uname)" == "Darwin" ]; then
+      open http://localhost:3000
+  elif [ "$(expr substr $(uname -s) 1 5)" == "Linux" ]; then
+      xdg-open http://localhost:3000
+  elif [ "$(expr substr $(uname -s) 1 10)" == "MINGW32_NT" ]; then
+      start http://localhost:3000
+  elif [ "$(expr substr $(uname -s) 1 10)" == "MINGW64_NT" ]; then
+      start http://localhost:3000
+  fi
+fi