+++ RELEASE_NOTES.md
@@ -4,6 +4,8 @@
 
 ### New features
 
+- Anonymize IP and email in logs
+
 ### Bug fixes
 
 - Change the keycloak config file to manage the token refresh
@@ -11,6 +13,7 @@
 ### Documentation
 
 - Improve architecture documentation
+- End to end demo including UI
 
 ### Continuous Integration